Added initial ports files.
This particular port is rather simple and straight forward. There
is no need for pkg-plist because there will only be one or two files
actually installed (once all is said and done). This initial port
is the starting point, and will be modified as the software it installs
is developed. Until Appy is ready to be released, this port is for
testing the work in progress. The remote server is fundametally the
most important part of the software, so it will be great to have this
point to start from.